refactor(rapid-mac): drop the in-app DMG installer, let Sparkle own updates
Sparkle has owned production updates since #1907: it runs the schedule,
downloads in the background, verifies the EdDSA signature and the
code-signing continuity, and installs on quit — including the authorization
path for a non-writable /Applications, which the in-tree installer never had.
The hand-rolled DMG pipeline it replaced stayed behind it as a fallback and
was still wired to live UI.

That fallback was not inert. Installer.defaultDownloader wrapped its progress
callback in withoutActuallyEscaping while URLSession holds its delegate
strongly and every didWriteData hop enqueues another @MainActor task holding
the same closure, so the runtime's escape check could still see a live
reference when the block returned and trap on brk #0x1. The crash lands in
UpdateInstallView.startInstall -> Installer.install(from:) with no message,
because the check compiles to a bare cond_fail.

Three of the four entry points already routed to Sparkle when it was
configured; FailedReplaceBanner did not, so its "Open update dialog" button
handed the crashing path to precisely the users whose manual install had just
failed. The invariant that the legacy path "no longer downloads or installs
production updates when Sparkle is configured" lived only in a comment, so a
fourth entry point could always drift back in. Deleting the installer makes
it structural.

- Delete Installer.swift, UpdateInstallView.swift and the update-install
  window scene, plus their tests.
- Route the tray item, the Settings CTA and the failed-replace banner to
  sparkle.checkForUpdates(); disable the affordances on unsigned builds,
  where Sparkle has no key and there is no longer anything to fall back to.
- Keep UpdateChecker as the read-only version/status source the pill and the
  Settings panel already render. Move updateDownloadHostAllowlist next to
  updateReleaseHostAllowlist; it now gates manifest shape, not a fetch.
- Drop the RAPID_GUI_UPDATE_CURRENT / RAPID_GUI_OPEN_UPDATE_WINDOW harness
  hooks and the golden-flow block that drove the deleted window.

Existing installs at 0.12.11 and older carry the trap in their own binary and
cannot be repaired by any release; they need one manual DMG download to reach
a Sparkle-enabled build.
